+++ src/st-table/index.ts
@@ -1,0 +1,1 @@
+export * from './st-table.module';