+++ src/st-table/shared/order.ts
@@ -1,0 +1,8 @@
+export enum ORDER_TYPE { ASC, DESC };
+
+export class Order {
+   constructor(
+      public orderBy: string,
+      public type: ORDER_TYPE
+   ) { }
+}