+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Extract theme of search
 * Reduce bundle size
 * Add Breadcrumbs component 
+* Create table component
 
 ## 1.4.1 (April 17, 2017)